begin SegmentedIter implementation
A SegmentedIter returns &[T] from the segments. This will be more efficient than accessing
single elements by index. Later this will be used to improve the Slice iterator (and others as
well). SegmentedIter is exported as public interface as well. Since this is SegVec there is
nothing to hide about that it is based on Segments.