Subslicing and SegVec::from(Slice) (#26)
* begin SegmentedIter implementation

A SegmentedIter returns &[T] from the segments. This will be more efficient than accessing
single elements by index. Later this will be used to improve the Slice iterator (and others as
well). SegmentedIter is exported as public interface as well. Since this is SegVec there is
nothing to hide about that it is based on Segments.

* new SliceIter implementation, using Flatten<SegmentedIter>

Added benchmark, improvement for iterating slices is +64%

* This magic #[inline] makes it 80% faster again!

* add benchmark comparing Iter vs SliceIter

* implement Slice::slice() -> Slice for subslicing

* implement From<Slice|&Slice> for SegVec

Creates a new/owned SegVec from a given slice

* Add benchmark: iterate over Vec

* FIX: indexing benchmark did modulo one where it should be 8900

* put test-vec generation into the bench-functions

* Remove the 'From<Slice>' impls

These will be reimplemented soon using from_iter() as in:
https://github.com/mccolljr/segvec/pull/26#issuecomment-1614107293